requirements/dev.txt
-r common.txt
django-seed==0.2.2
Faker==4.0.0
coverage==5.2.1
factory-boy==2.12.0
